Reorganize fitting scripts and data
