--- test_fitting_data/210526_m04_p06_V_A_.nd2_dt.txt
@@ -1,1 +1,0 @@
-   1.0006004e-02   2.0012007e-02   3.0018011e-02   4.0024014e-02   5.0030018e-02   6.0036022e-02   7.0042025e-02   8.0048029e-02   9.0054032e-02   1.0006004e-01   1.1006604e-01   1.2007204e-01   1.3007805e-01   1.5009005e-01   1.6009606e-01   1.8010806e-01   1.9011407e-01   2.1012608e-01   2.4014409e-01   2.6015609e-01   2.9017410e-01   3.1018611e-01   3.5021013e-01   3.8022814e-01   4.2025215e-01   4.6027617e-01   5.1030618e-01   5.6033620e-01   6.1036622e-01   6.7040224e-01   7.4044427e-01   8.1048629e-01   9.0054032e-01   9.8058835e-01   1.0806484e+00   1.1907144e+00   1.3107865e+00   1.4408645e+00   1.5909546e+00   1.7410446e+00   1.9211527e+00   2.1112668e+00   2.3213928e+00   2.5515309e+00   2.8116870e+00   3.0918551e+00   3.4020412e+00   3.7422453e+00   4.1124675e+00   4.5327196e+00   4.9829898e+00   5.4832900e+00   6.0236142e+00   6.6339804e+00   7.2943766e+00   8.0248149e+00   8.8252952e+00   9.7058235e+00   1.0676406e+01   1.1747048e+01   1.2917751e+01   1.4208525e+01   1.5629378e+01   1.7200320e+01   1.8921353e+01   2.0812487e+01   2.2893736e+01   2.5175105e+01   2.7696618e+01   3.0468281e+01   3.3510106e+01   3.6862117e+01   4.0554333e+01
