--- test_fitting_data/210526_m04_p06_V_A_.nd2_q.txt
@@ -1,1 +1,0 @@
-9.4398818e-02   1.8879764e-01   2.8319645e-01   3.7759527e-01   4.7199409e-01   5.6639291e-01   6.6079172e-01   7.5519054e-01   8.4958936e-01   9.4398818e-01   1.0383870e+00   1.1327858e+00   1.2271846e+00   1.3215834e+00   1.4159823e+00   1.5103811e+00   1.6047799e+00   1.6991787e+00   1.7935775e+00   1.8879764e+00   1.9823752e+00   2.0767740e+00   2.1711728e+00   2.2655716e+00   2.3599704e+00   2.4543693e+00   2.5487681e+00   2.6431669e+00   2.7375657e+00   2.8319645e+00   2.9263633e+00   3.0207622e+00   3.1151610e+00   3.2095598e+00   3.3039586e+00   3.3983574e+00   3.4927563e+00   3.5871551e+00   3.6815539e+00   3.7759527e+00   3.8703515e+00   3.9647503e+00   4.0591492e+00   4.1535480e+00   4.2479468e+00   4.3423456e+00   4.4367444e+00   4.5311433e+00   4.6255421e+00   4.7199409e+00   4.8143397e+00   4.9087385e+00   5.0031373e+00   5.0975362e+00   5.1919350e+00   5.2863338e+00   5.3807326e+00   5.4751314e+00   5.5695302e+00   5.6639291e+00   5.7583279e+00   5.8527267e+00   5.9471255e+00   6.0415243e+00   6.1359232e+00   6.2303220e+00   6.3247208e+00   6.4191196e+00   6.5135184e+00   6.6079172e+00   6.7023161e+00   6.7967149e+00   6.8911137e+00   6.9855125e+00   7.0799113e+00   7.1743101e+00   7.2687090e+00   7.3631078e+00   7.4575066e+00   7.5519054e+00   7.6463042e+00   7.7407031e+00   7.8351019e+00   7.9295007e+00   8.0238995e+00   8.1182983e+00   8.2126971e+00   8.3070960e+00   8.4014948e+00   8.4958936e+00   8.5902924e+00   8.6846912e+00   8.7790900e+00   8.8734889e+00   8.9678877e+00   9.0622865e+00   9.1566853e+00   9.2510841e+00   9.3454830e+00   9.4398818e+00   9.5342806e+00   9.6286794e+00   9.7230782e+00   9.8174770e+00   9.9118759e+00   1.0006275e+01   1.0100673e+01   1.0195072e+01   1.0289471e+01   1.0383870e+01   1.0478269e+01   1.0572668e+01   1.0667066e+01   1.0761465e+01   1.0855864e+01   1.0950263e+01   1.1044662e+01   1.1139060e+01   1.1233459e+01   1.1327858e+01   1.1422257e+01   1.1516656e+01   1.1611055e+01   1.1705453e+01   1.1799852e+01   1.1894251e+01   1.1988650e+01   1.2083049e+01   1.2177447e+01   1.2271846e+01   1.2366245e+01   1.2460644e+01   1.2555043e+01   1.2649442e+01   1.2743840e+01   1.2838239e+01   1.2932638e+01   1.3027037e+01   1.3121436e+01   1.3215834e+01   1.3310233e+01   1.3404632e+01   1.3499031e+01   1.3593430e+01   1.3687829e+01   1.3782227e+01   1.3876626e+01   1.3971025e+01   1.4065424e+01   1.4159823e+01   1.4254221e+01   1.4348620e+01   1.4443019e+01   1.4537418e+01   1.4631817e+01   1.4726216e+01   1.4820614e+01   1.4915013e+01   1.5009412e+01   1.5103811e+01   1.5198210e+01   1.5292608e+01   1.5387007e+01   1.5481406e+01   1.5575805e+01   1.5670204e+01   1.5764603e+01   1.5859001e+01   1.5953400e+01   1.6047799e+01   1.6142198e+01   1.6236597e+01   1.6330995e+01   1.6425394e+01   1.6519793e+01   1.6614192e+01   1.6708591e+01   1.6802990e+01   1.6897388e+01   1.6991787e+01   1.7086186e+01   1.7180585e+01   1.7274984e+01   1.7369382e+01   1.7463781e+01   1.7558180e+01   1.7652579e+01   1.7746978e+01   1.7841377e+01   1.7935775e+01   1.8030174e+01   1.8124573e+01   1.8218972e+01   1.8313371e+01   1.8407769e+01   1.8502168e+01   1.8596567e+01   1.8690966e+01   1.8785365e+01   1.8879764e+01   1.8974162e+01   1.9068561e+01   1.9162960e+01   1.9257359e+01   1.9351758e+01   1.9446156e+01   1.9540555e+01   1.9634954e+01   1.9729353e+01   1.9823752e+01   1.9918151e+01   2.0012549e+01   2.0106948e+01   2.0201347e+01   2.0295746e+01   2.0390145e+01   2.0484543e+01   2.0578942e+01   2.0673341e+01   2.0767740e+01   2.0862139e+01   2.0956538e+01   2.1050936e+01   2.1145335e+01   2.1239734e+01   2.1334133e+01   2.1428532e+01   2.1522930e+01   2.1617329e+01   2.1711728e+01   2.1806127e+01   2.1900526e+01   2.1994925e+01   2.2089323e+01   2.2183722e+01   2.2278121e+01   2.2372520e+01   2.2466919e+01   2.2561317e+01   2.2655716e+01   2.2750115e+01   2.2844514e+01   2.2938913e+01   2.3033312e+01   2.3127710e+01   2.3222109e+01   2.3316508e+01   2.3410907e+01   2.3505306e+01   2.3599704e+01   2.3694103e+01   2.3788502e+01   2.3882901e+01   2.3977300e+01   2.4071699e+01   2.4166097e+01
